Simplify requires.
